+++ Source/CourseDashboardViewController.swift
@@ -282,3 +282,25 @@ extension CourseDashboardViewController {
         return tabBarItems
     }
 }
+
+extension UITabBarController {
+    
+    func tabBarViewControllerIndex(with controller: AnyClass, courseOutlineMode: CourseOutlineMode? = .full) -> Int {
+        guard let viewControllers = viewControllers else {
+            return 0
+        }
+        
+        for i in 0..<viewControllers.count {
+            if viewControllers[i].isKind(of: controller) {
+                if  viewControllers[i].isKind(of: CourseOutlineViewController.self)  {
+                    if let viewController = viewControllers[i] as? CourseOutlineViewController, viewController.courseOutlineMode == courseOutlineMode {
+                        return i
+                    }
+                } else {
+                    return i
+                }
+            }
+        }
+        return 0
+    }
+}
